+++ changelog.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Ability to install modules from FORGEBOX that have secure storage.
+- If a module's box.json exposes a `boxlang.executable` property, create a symlink to that executable in the BoxLang home bin folder.
+- If a module's box.json exposes a `boxlang.executables` property, create those shell scripts in the BoxLang home bin folder.
 
 ### Fixed